Worker memory optimization for 12gb vram

+++ worker/sd_inpaint_model.py
@@ -60,7 +60,7 @@ def initialize_model(config, ckpt):
     model.load_state_dict(torch.load(ckpt)["state_dict"], strict=False)
 
     device = torch.device("cuda") if torch.cuda.is_available() else torch.device("cpu")
-    model = model.to(device)
+    model = model.to(torch.float16).to(device)
     sampler = DDIMSampler(model)
 
     return sampler
