+++ worker/sd_text2im_model.py
@@ -140,7 +140,7 @@ class StableDiffusionText2ImageModel:
     def __init__(self):
         args = _default_args
         self.device = torch.device("cuda")
-        self.model = load_model()
+        self.model = load_model().to(torch.float16)
         self.model = self.model.to(self.device)
         if args.plms:
             raise NotImplementedError("PLMS sampler not (yet) supported")
